kola/tests: move GetRpmOstreeStatusJSON to util package
The func is getting used in multiple places, so make it part of the
`util` package.  The tests that use the func have been updated
accordingly.
